[IMP] l10n_mx_edi: New changes to payment complement.
The SAT was improved the payment complement.

Previously they had a restriction that if the invoice currency is MXN
the rate must be 1.

But now was changed, and always must be assigned the rate in the payment
currency.

With this We avoid stamp errors.

Also with this We do not need change the payment total.

closes odoo/enterprise#2926