[FIX] l10n_co_edi: Invalid XML ID in view
The previous xml id was referencing "product.template_..." which
was causing trouble when upgrading other modules. This PR correct
xml id to the regular "template_...".

Steps to reproduce:
1) Install accounting
2) Install "Electronic invoicing for Colombia with Carvajal"
3) Upgrade accounting

Caution ! If the columbia l10n was installed already, we have to
manually remove the two previous views. We tried some workaround
to have it removed automatically but it didn't went well.

closes odoo/enterprise#2923