[FIX] account_reports: fixes in partner ledger
This commit includes several fixes on the partner ledger report:
1) The button to open the partner ledger on res.partner was calling the report which was filtering on the partner only thanks to the 'partner_ids' in context (see _query_get on account.move.line)... although there is a 'line_id' parameter that does explicitly the job.
Using this parameter instead,
* allows to remove the useless total line when there's only one partner
* is a cleaner implementation of the partner search

2) When selecting a partner in the filter, it wasn't unfolded by default. We now unfold automatically as soon as there's only one partner displayed

3) We aslo replace the customly made select2 filter with real standalone m2mtags widget.
Partner filter was introduce at dfa1d58 and used a server-side search() on all res.partner entries to build the dropdown list of partners. This implementation crashes the server with a MemoryError if there is a lot of records in the res_partner table.
To handle this issue ina cleaner way, we just replace the select2 filter with m2mtags widget.

This was asked in task: 36626
Closes #2680

Co-authored by qdp-odoo@openerp.com

closes odoo/enterprise#2952