[FIX] account_reports: Support special chars in account footnotes
A traceback was raised when using a special char
e.g. "crédit"
in a footnote to an account in the ledger
(Open the ledger, dropdown the account, add a footnote)

This is because the string was sent not encoded in UTF-8
from the web client to the server.

The server now ensures to pass the string UTF-8 encoded
to the render

opw-1903439