[FIX] web_studio: properly remove dropdown from kanban
Previously, when user tried to remove an existing dropdown from kanban using
Studio, it used to remove the whole kanban container instead of particular
dropdown node.

This issue was caused due to passing improper node to the `node_clicked` handler.
We used to manually create dropdown node object by providing just tag name and
attrs, and pass it to the handler. And because of this, we were not able to find
correct `xpath_info` using `findParentsPositions` method of ViewEditorManager.

This commit fixes the issue by finding the actual node from the arch and pass
it to the handler.

This commit also adds `name` attribute to new dropdowns being added using
Studio. It will prevent Studio from adding random name attribute while
normalizing the arch. If name attribute is not provided by default, normalize
method will add different name attributes, firstly while adding a new dropdown,
and again while removing it (without saving the changes). So it will result into
warning of node not being located.

Task 1866571

closes odoo/enterprise#2960