[IMP] l10n_mx_edi_payment_bank: Added fields to save the account used to the payment
The CFDI document that is generated in each payment has the next
attributes:

- RfcEmisorCtaOrd: Is the NIF of the entity from where was made the
  payment. If this entity is foreign must be used the generic value
  XEXX010101000.
- NomBancoOrdExt: Is the name of the entity from where was made the
  payment. If the entity is foreing this value must be established.
- CtaOrdenante: Is the account number that was used in the payment.o
- RfcEmisorCtaBen: Is the NIF of the company bank where was received the
  payment.
- CtaBeneficiario: Is the company bank account where was received the
  payment.

To this was added the next fields:
- VAT in the bank to has this value by each bank.
- Partner bank in the account.payment and in the wizard that allow make
the payment by many invoices to allow set the bank account that was used
the partner to make the payment.

This new attributes was added in the PDF, in the l10n_mx_edi because
that attributes exist on the template but are not filling now.

closes odoo/enterprise#2974