[FIX] account_budget: properly calculate percentage sum in read_group
The percentage field is computed in _compute_percentage as a value
between 0 and 1. This is good because that's what widget="percentage"
in views expects.

This makes read_group also return a value between 0 and 1, instead of
0 and 100. Otherwise list views will show e.g. 5000% when grouping
lines that sum to 50%.

Introduced in saas-11.4 at odoo/odoo@9148960174bdb0abbee6a5b7403a8f.

opw-1904481

closes odoo/enterprise#2975