[FIX] account_bank_statement_import_ofx: set currency for import ofx test
In the revision
odoo/enterprise@2df1fe8cc13a171ae97f77fe16cb97942f282dce

Setting the journal currency to USD has been removed
because of the revision
odoo/odoo@22f0da809e84618901b7ca4b78582abd64c03c55

Which prevents to set the currency of a journal
to the company currency, as it must not happen.

According to the db state, the company is not always
`base.USD`, and so just simply skipping to set
the company currency to USD is not enough.

The test should check the company currency,
and if it's not `base.USD`, it must set it,
as the bank statement imported is within USD.

opw-1889980