[IMP] account_sepa_direct_debit: group payments by journal instead of mandate in XML file
Before, we created a distinct PmtInf node by mandate, each of them containing the payments made thanks to it as transactions. Though it was 100% compliant with the SEPA regulation, it caused some problems in the Netherlands, where some banks limit the number of PmtInf nodes allowed in the file. This commit fixes that, by reducing the number of PmtInfo nodes created to the minimum, that is, 1 per creditor account (hence, one per payment journal).

closes odoo/enterprise#2981