[FIX] web_enterprise: increase cross click area in burger menu
Before this commit, it was difficult to click on the cross
to close the burger menu because the clicking area was very
small.
Even worse, if the user clicked next to the zone, the burger
menu app was opened instead of closing the burger menu.
It was very confusing.

Now, we removed some paddings/margins on the top to give
more space to the cross area zone.

Task ID: 1893168

closes odoo/enterprise#3024