[FIX] stock_account_enterprise: avoid to change returned value in read_group
Before this commit, 'read_group' on 'stock.report' returned an unexpected value.
When there was no record, [{}] was returned instead of [].
This resulted in side effects like in graph view.

Steps to reproduce:
without demo data
- Go to inventory-> Warehouse Dashboard
- Then click on timerange-> "transfer date" and compare to "Previous Period".
- Move mouse on chart
-> Traceback: Uncaught TypeError: Cannot read property 'x' of undefined

It was not necessary to override the returned value.
So, we simply adapted the code to make sure that the resulted value is not
empty before updating it (res[0].update[..]).

Taks-ID: 1886908

closes odoo/enterprise#3017