[FIX] account_reports, l10n_es_reports: make from_context report lines work when printing reports
Before that, a financial report line defined with balance=from_context had always a balance of 0 when exporting the report to file. This was due to the fact the download of the exported file is made using an http route, which destroys the context. Now, we use the options to store these values, so that they persist when calling the route.

closes odoo/enterprise#3026