[IMP] IoT : IoT refactoring - With longpolling
Improve the implementation of IoT.
Add distinct driver for : Camera and Printer
Allow both-direction communication between iot box and devices.

For the longpolling you need to create a listener in witch you have a unique session_id
and a list of device_id that you want listen.
And you can make a action request.

Depending on the devices of the box we do not know when the result of the query will arrive
so it will be returned in the result of the route "event" made with the listener

task : 1906910