[FIX] l10n_co_edi: handle Carvajal's flaky API
It happens occasionally that the upload method times out. This throws
an ugly traceback which would be ok, except that this timeout does not
necessarily mean that Carvajal didn't receive the invoice. Cases have
been observed where the upload timed out, but Carvajal still committed
the invoice on their side. This is problematic because the traceback
will cause Odoo to roll back the transaction and thus not assign the
current sequence value which was was sent to Carvajal. When another
invoice gets validated afterwards you end up with 2 different invoices
on Carvajal's side with the same name.

To resolve this, catch this error and log a note on the invoice. This
way the invoice still gets validated and assigned the sequence sent to
Carvajal. At a later point the user can reupload the invoice to
Carvajal which is handled correctly, even if the invoice already
existed.

closes odoo/enterprise#3044