[FIX] web_dashboard: reduce margins for better visibility
(1) Before this commit, there was a lot of empty space in
dashboard views.
For now, it's not possible to reduce margins to 0 because
otherwise the linechart will be stick to the next group.
A specific css rule to only apply bottom margin for line chart
is not possible as we can't distinguish chart types.
This solves some issues in mobile and make it usable.

(2) We also override border width to make it more thicker in
dashboard views.

Task-1889258

closes odoo/enterprise#3046