[FIX] web_[enterprise,studio]: make runbot green again
There were (at least) two reasons why the studio 'tests' tour was failing under
undeterministic circumstances.

The first one was after the new app creation: the tour was not waiting for the new
app to be fully ready (form editor rendered in the dom and toggle menu ready). A
very old rev. (see odoo/odoo@af3dbd5), kept during all these years through all
refactorings, was resolving the 'app_clicked' deferred after an arbitrary 2sec timeout.
If the Studio app creation was taking more than 2sec, the home menu was toggled an the
next step was wrongly run. Also note that the `o_menu_toggle` step has been improved
because one can click on the menu toggle after reinstatiation even though the icon is
not yet toggled.

The second one was after a new menu creation: the tour was not waiting for the new
menu to be fully ready (menus and edit menu modal updated). The first modal (new menu
creation) was closed (so the second modal could be closed) but the second modal was
reloaded with the new menu so the modal was still present (which was preventing the
next step to be executed). Now, the first modal is not closed directly (it will be
destroyed and reinstantiated at reload).

Complement to odoo/enterprise@584d319