[FIX] web_studio: reload the page in Studio
Since rev. odoo/enterprise@3f16b50 it wasn't possible to reload the page
while editing an action with Studio (a traceback was occuring then).

While instantiating menu widgets in `show_application`, the studio menu
was rendered without any action (because no action had already been loaded).

The issue here is that the `studioMode` is set too soon (before the
`show_application`). When reloading the page, Studio will be reloaded entirely
so no need to eagerly render the Studio menu.

It's (sadly) not possible to test this feature however because reloading the
page would stop the current tour execution.

Closes #3078