[FIX] account_reports: long account names
The 'sheet' view was originally removed in order to be able to display
several comparison columns in the Trial Balance with full account names.
However, the final result is not easily readable.

Therefore, we truncate the long account names and re-introduce the
'sheet' view.

Work done by CSN in master branch, partially cherry-picked in v12.

Co-authored-by: Cedric Snauwaert <csn@openerp.com>

closes odoo/enterprise#3080