[FIX] marketing_automation: Filter on resource_ref doesn't work
Creating a search filter on field resource_ref cannot work as this is non stored reference field.
So the only way to make it working would be to fetch all the records of model 'marketing.participant'
and test which one would be the one with the reference name written in the search bar.
But it's completely not optimal.

opw:1889713

closes odoo/enterprise#3084