[FIX] web_studio: use Studio bundle in tests
Since rev. odoo/enterprise@bead68c the Studio bundle is not always loaded in report assets
but conditionally if `studio` is in the report rendering values.

However, the helper function that loads css for report editor tests was not setting `studio`
so the Studio bundle was not in tests, which makes *some* tests fail on *some* browsers
(runbot was not suffering from this because the tests were passing in chrome headless).

closes odoo/enterprise#3091