[FIX] hr_appraisal: restrict survey user rights to manager
Survey users can design, share and tests all surveys by default.
That would be fine, except that the appraisal module grants
the Survey: User group to all employees.
This privilege is given them only to have them be able to answer and review
their answers to the surveys.
So short of creating a new group for that purpose, we hide these functionalities
for users that are not survey manager.

We also restrict creation, edition and deletion of surveys to managers.

opw 1896360

closes odoo/enterprise#3092