[FIX] sale_subscription: make `in_progress` read-only
Salespeople may not have the right to alter subscription stages.
However, since `in_progress` field is related to the stage (and thus
dynamically updated by onchange calls), the form view is going
to save its value as well whenever the `stage_id` of a subscription
changes (as long as the field is present in the view).

Making it read-only prevents saving this redundant value, and avoids
triggering an AccessError for users with no write access to stages.