[FIX] sign: Set template uploader as sender when sharing a document
Purpose
=======

When sending a document to sign by using the 'Share' button, the sign.request
that is created is created with sudo(). That way, the sender is named 'Odoo Bot'
which is not very sexy.

Specification
=============

Use sudo(template.create_uid) to create the sign request. That way, this is the
person who uploaded the document that creates the request (eg: a vendor, a marketing
officer,...).

closes odoo/enterprise#3110