[FIX] account_sepa_direct_debit: use BtchBookg node in XML generated from batch payments
Use the batch booking option of SEPA Direct debit properly with the new implementation of the module, using account.batch.payment objects. This option has to be enabled for everyone by default (both old and new dbs), as it the reason batch payments exist. It can however be deactivated, setting a account_sepa_direct_debit_no_batch_booking config parameter.

//NOTE TO FORWARD PORT:
This commit is actually just the opposite of https://github.com/odoo/enterprise/pull/3111/commits/4ff8f2738050da50ea8c624fc9a75f76805a1185 , in v11 (so please do not overwrite it with the v11 one). The reason for that is that we want to activate batch booking even if the config parameter is not set (for already existing dbs, so).

closes odoo/enterprise#3112