[FIX] web_studio: use existing _view_obj method to retrieve view
Normally, calling get_template() with an xmlid (string) ends up calling
get_view_id() that is overriden in website to return the more specific view if
the xml_id is a string (and not an ID).

But since studio override get_template() and return an ID no matter what, it
ends up breaking the multi-website feature since get_view_id() always get an ID
instead of the string xml_id that would allow to find the more specific view.

With this commit:
We use the already existing _view_obj() instead of searching directly the view.
That method is basically doing the same except it does not search only 'qweb'
views.
That method is overriden in website to find the most specific one.
Using it in studio will then keep the multi-website behavior when both modules
are installed.

opw-1903315
opw-1902766

closes odoo/enterprise#3126