[FIX] web_studio: type of filename binary instead of char
When creating a file field using studio, two files are needed,
the first one type binary for the file content and the second
type char for the filename.

The problem is the internal json key used to defined the key
of the type of the filename field was `ttype` instead of `type`.
The creation of the field used a wrong `binary` type (set by
default with the `type` key) instead of the one in the `ttype`.

Related to 6d29a723da51560b

opw-1902334

closes odoo/enterprise#3135