[FIX] web_studio: duplicate report view inheritance
When we duplicate a report in studio, this will duplicate the "main"
view and all the viewed t-called recursively.

We duplicate these views will all inheritance applied: so if view A
is modified by views B, D, E the only created view is the combination
of A + B + D + E.

But there was an issue when we duplicate A that modify B, we would
create a combinated view of A + B but they were still set to modify B
which makes no sense since the combinated view is B extended.

Added test without change fails with:

```
  File ".../test_report_editor.py", line 50, in test_copy_inherit_report
    report.copy_report_and_template()
  ...
  odoo.exceptions.ValidationError:
    Error while validating constraint
    '<t t-name="web_studio.test_inherit_report_user_copy_1">' cannot be
    located in parent view
```

opw-1905837
closes #3140