[FIX] project_forecast: clean context when accessing the forecast grid view
Suppose you have added via studio a custom field on project.project,
and you use that field to group records.
Then go on a record, and try to access the project forecast.
It results on a traceback.
What hapens is that view_monthly_forecast loads the forecast grid view,
passsing the context it got from the view on project.project.
Since the grid view uses what is in the context with the hightest priority
to determine the group_by, it crashes since the field is unknown on that model.

We simply pop that value from the context.

Note that if the same groupby is selected by the user as default filter,
then there is no bug since the group_by is not added to the context in that case
(which is consistent with dropping it from context)

opw 1906669

11.0-opw1906669-grid-len

closes odoo/enterprise#3147