[FIX] website_sale_dashboard: field not defined in module
The `days_to_confirm` field of the `sale.report` model is defined
in the `sale_enterprise` module which is not in the dependencies
of the manifest file of `website_sale_dashboard`.

Thus uninstalling the `sale_enterprise` module remove the
required field and accessing the website results in a traceback.

This commit remove the line in the template that was using that
field.

Related to bd3d4496762f0b8

opw-1909470

closes odoo/enterprise#3154