[IMP] l10n_mx_edi: Use strip in method that get serie/folio
In the CFDI generation, when a element is empty this does not must be
added, By this reason is used the None value.

But if a element is ' ', is not the same that None.

Then, when the invoice name do not have prefix, the serie is empty, but
when you call that method in the CFDI payment generation, that is not
None and try add that attribute in the CFDI, and like that cannot be
empty return the error:

El CFDI no tiene una estructura XML correcta. Folio: 8404. Serie: CUST
IN 2018. Exception in startElement: cvc-pattern-valid: Value '' is not
facet-valid with respect to pattern '[^|]{1,25}' for type
'#AnonType_SerieDoctoRelacionadoPagoPagos'.

closes odoo/enterprise#3163