[IMP] l10n_mx_reports: First DPIVA version
DPIVA is an alternative that have the contributor to present your TAX
declaration for your operations made with thirds. Its similar that the
DIOT report, but some companies need to generate the DPIVA.

This is a txt file that is send to a Digital Document Reception Service
Provider (PSRDD), (similar to a PAC). And that is send to the SAT.

Then We need to generate a TXT file with the structure provided and send
it.

How was resolved?

Was added a new button on the DIOT report, where you can download the
TXT file with the same information that the DIOT is generated but with
the new structure.

closes odoo/enterprise#3173