[FIX] delivery_fedex: get label without sale order
Since the fix in rev: https://github.com/odoo/enterprise/commit/d68c1e6c03887c0d087eba14360fdb172f1c0546 it is impossible to get a shipping label without
wihtout having a link to the sale order.

We fix it by taking the product price if the operation is not linked to
a sale order line through its move, and we use the default price of the
product if lines are not linked to any sale order.

closes odoo/enterprise#3194