[FIX] web_studio: correctly set context when reloading action
In Studio, multiple user interactions require to reload the action being edited
(i.e. adding/removing a view type). Before this rev. there was a traceback when
doing so for actions using `active_id` in their context.

This is because the action preprocessing wasn't aware of the additional context
(i.e. `active_id`) normally set by the ActWindowActionManager when loading the
action.

opw-1904361
Closes #3103