[FIX] account_batch_payment: relax type check
As of version 12.0 and the refactoring of batch payments via
9ebfd5acae5aa7b2afbb8a3c5b83d24c8fba0057, constraints on batch payment
had been reinforced, including a check for uniform payment types
inside a batch (inbound/outbound).

This constraint was a bit too restrictive. For example a batch may
represent a payout from a payment acquirer, which will often contain a
majority of inbound payments, but could have a few "outbound" payments
in the mix (direct refunds or chargebacks).

The batch itself does have a main batch_type (inbound/outbound) driving
its name and sequence, but it is possible and supported to have a few
reversed payments in the batch. For this reason, it makes sense to keep
the constraint requiring matching batch_type and (some) payment_type.

closes odoo/enterprise#3198