[FIX] web_grid: button click no affect next click
When clicking on a action button in the grid view, the action will get
the compounded context of the current model overridden by the current
button.

But this compound context was set on the current button, so an out of
date context would possibly override the up-to-date context.

eg. on the timesheet we could click validate with anchor_date 2018-01-31
    then without reloading, the anchor_date would forever be 2018-01-31
    even after changing week or month.

The added test without code change failed with:

  4. second button click get current grid anchor date
     (expected: 2017-01-24,2017-01-17, result: 2017-01-24,2017-01-24)

opw-1915033
closes #3214