[FIX] account_sepa_direct_debit: allow registering a SDD payment without going through an invoice's "register payment" button
Before that, even when a mandate was available, directly creating a SEPA Direct Debit payment (hence, not from the "register payment" button of invoices) did not register it on the mandate, which ultimately made it impossible to generate any xml data for this payment, and thus actually collect it.

Now, we do the payment registration on mandate and the mandate check in a more generic place, so that it is always called, and not only from "register payment".

closes odoo/enterprise#3215