[IMP] l10n_mx_edi: Improved way to get the refunds related.
When a invoice is reconciled with a refund, that information is added on
the invoice payment data, and that is not consider like a payment in the
CFDI for the payment complement.

But in the next case:
1. Create an invoice for 100 USD
2. From the `Invoice Refund` menu create a new invoice refund to the
same partner that in the invoice in the step 1.
3. Reconcile both records created on step 1 and step 2.

In this case Odoo do not relate the invoice with your refund, then with
the last code sum the refund amount with the amount paid.

By this reason, now before to search the lines I check that the move
related has not a invoice related (the refund) or if this attribute is
fill, then check that is different to my invoice.

Based on
[this](https://github.com/odoo/odoo/blob/12.0/addons/account/models/account_invoice.py#L208)

closes odoo/enterprise#3216