[FIX] sale_subscription: impossible to delete subscription
Deleting a subscription with some description lines resulted in a
AccessError due to a field in subscription removed before
accessing the subscription_lines which has a access rule depending
on that removed field.

This PR surcharges unlink to remove the subscription lines first
and then remove the subscription itself.

opw-1913748

closes odoo/enterprise#3218