[FIX] web_studio: no traceback on reset background
Use case: open studio on home menu and reset background -> studio is
closed and there is a traceback, a traceback at each F5 and when opening
studio.

This is because we do the "reload" action, and a code in this instance
try to toggle off the home menu which makes no sense in this instance.

This does not happen when setting the background since in this instance
we just do window.location.reload().

opw-1915223
closes #3231