[FIX] web_enterprise: home menu allow ctrl+c/x
When doing copy/cut (CTRL+C/CTRL+X or COMMAND+C/COMMAND+X in apple) the
home menu will focus inside the app search bar, thus it is not possible
to do copy from the home menu of something (a chat window, something in
the preferences modal, ...).

With this changeset these commands can be done.

opw-1916024
closes #3240