[FIX] web_studio: prevent issue with rename and slow network
There was an issue with the field renaming in Studio: with a slow network (or
very quick reflexes), it was possible to drag and drop something (another field
for example) next to a field being renamed.

The blockUI was removed after the rename RPC, not after the view re-rendering so
the hook was still set with the old field name (which was sometimes make the
tour crashes).

Morever, the rename rpc was executed in the MutexedDropPrevious instance, which
does not make any sense because we don't want to drop it (only the view editions
are meant to be executed in the MDP).

Note that the use of `always` has been avoided (then and fail instead) in
prevision of the jQuery new version.

Task 1916053

closes odoo/enterprise#3254