[FIX] marketing_automation: flatline on ActivityGraph opw-1912452
The ActivityGraph in the marketing automation campaign was always
showing a flatline.

Now the ActivityGraph shows the number of emails success and rejected.

Due to the refactoring of Date/Datetime fields in v12, now in the date
fields are stocked in the database in a datetime type and not in string
type as it was done until now.

closes odoo/enterprise#3278