[FIX] web_enterprise: fix border color of .card
The migration to BS4 introduced card instead of panel and well, and we kept the
border that was previously on the well. However this is not necessary for card.

This commit will let the card border take its default relevant value instead of
a white that is invisible on our usual white background. Before this commit the
border of the card also didn't adapt depending on the background color.