[FIX] timesheet_grid_sale: apply method on correct model
In community, there is a method returning the domain to
restrinct the timesheet to link to the invoice, when
invoicing from a sales order. This method is declared on
account.invoice.line. In enterprise, this method should
be extended to custom the domain with the validation.
Problem: the extention is declared on account.invoice and
not account.invoice.line. So the override is never called.
This commit fixes the issue by changing the model name (as
there is only this method in the file).

closes odoo/enterprise#3201