[IMP] l10n_us_check_printing: allow using US check format for other countries
Before that, this check format was only available for companies whose country was set to US. It made it impossible to use account_check_printing for companies in countries without any check layout (that is, everyone except US and Canada), even if they have a localization check layout module installed. To solve that, we introduce a config parameter named account_check_printing_force_us_format to force the use of the US layout (it thus still requires l10n_cus_check_printing to be installed).

closes odoo/enterprise#3282