[FIX] marketing_automation: fix editing of child activities
Backport of: c7c40bf01328455ba90699982a8efb2d82cbd790

To implement a hierarchical one2many, this modules overrides the one2many field.
It builds a tree based on a simple parent relation defined on the activities.
The root is thus the one2many field itself, which contains only the top-level
activities, and not activities that are children of other activities.

However when an activity is edited, we look to see if the records ID is already
in the one2many. Since it's not, we don't generate an UPDATE but an ADD command;
as a result nothing is done, since the ID really is in the one2many.
Thus the changes made are completely discarded on save.

opw 1892091

opw 1946673

closes odoo/enterprise#3793

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>