[FIX] marketing_automation: fix editing of child activities
To implement a hierarchical one2many, this modules overrides the one2many field.
It builds a tree based on a simple parent relation defined on the activities.
The root is thus the one2many field itself, which contains only the top-level
activities, and not activities that are children of other activities.

However when an activity is edited, we look to see if the records ID is already
in the one2many. Since it's not, we don't generate an UPDATE but an ADD command;
as a result nothing is done, since the ID really is in the one2many.
Thus the changes made are completely discarded on save.

opw 1892091

closes odoo/enterprise#3289