[IMP] web_enterprise: re-enable mobile_menu_tests
The mobile menu tests (aka BurgerMenu tests) were disabled due to
missing dedicated mobile test suite. As we now have this suite we can
re-enable them.

Note that the systray menu items should be limited to always-installed
modules to prevent conditional mocking of mail_service depending on
the presence of the "mail" module (cf. "mail" module is not in
auto-install mode).