[FIX] web_enterprise: Burger menu doesn't show menuitems w/o children
The inNavbar flag allow to differenciate navbar and dropdown items (aka
sub-menu items).

In mobile, every menu items are rendered as "in the (burger)menu/navbar" (and
not in a dropdown). We can then safely force this value to "true" to
prevent items not being displayed.

Bug introduced during Bootstrap 4 migration.

Related to task 1891955

closes odoo/enterprise#3253