[FIX] stock_barcode: location's fields in read
get_barcode_view_state returns the location for a picking or an
inventory. The location's fields are returned by a read() without
specific fields. By default all the fields present on the stock.location
object are returned including all the quants contained in those
location. The number of quants is potentially big and the only field
required by the client action are the id and the display_name of the
location.

This commit add only the useful fields in the read call.

closes odoo/enterprise#3309