[FIX] l10n_mx_edi_external_trade: fix installation of ir_model_data
Install the l10n_mx_edi_external_trade module, then uninstall it.
It is then not possible to reinstall it, because it created some ir_model_data
in l10n_mx_edi that is not uninstalled.
Since it tries to reinsert that data, this creates an integrity error.

Since this data is inserted in the data of another module (l10n_mx_edi)
we do not remove it at uninstall, as we could inadvertently remove some other
data, and leave the database in an inconsistent state.
So this leaves out the uninstall_hook, thus we prefer to simply check if the
data is not already present and only add it then.

Note that we also add a removal of duplicate cities that have been added with
l10n_mx_edi_code, as they may have already been added (without that code).

opw 1916374

closes odoo/enterprise#3341