[FIX] marketing_automation: fix crash on unlinked participants
Fine-tuning of commit d663e1ef1c40d8db30886980d128be17fedd9702
which refactored the unique field filtering to make it work, but introduced
(a?) new bug(s?) of its own.
We filter the records with a call to exists, and put values in a set instead of
a list to reduce the complexity.

A test has been added to avoid such regressions.

Special thanks to @wtaferner who reported the issue.

Closes #3321
opw 1892091

closes odoo/enterprise#3339