[IMP] l10n_mx_edi: Support cases with unreconciled lines
According to the quantity of support necessary, the last weeks about
[this improve](
https://github.com/odoo/enterprise/commit/e3d3176aa9f67d2b706f2baa5ebb202cc01c2655
), We improve the messages returned in the payment when the CFDI is not
generated where is indicated the reason that provoked that decision.

And related to [this Issue](https://github.com/odoo/odoo/issues/29446),
now is supported the case with multicurrency where the payment has
exchange rate difference and the journal entry is not reconciled.

closes odoo/enterprise#3357