[FIX] timesheet_grid: consistent model
In some cases, a non-manager might receive notifications to approve
timesheets, although they should only be sent to managers.

Timesheet encoding and approval reminders are handled by the same method
`_cron_timesheet_send_reminder`. This method sends the mail to the
corresponding employee at
https://github.com/odoo/enterprise/blob/fee4740b69a7e247a8c7708073abd523fc42654f/timesheet_grid/models/res_company.py#L162

An employee ID is used when rendering the email; however,
`mail_template_timesheet_reminder_user` and
`mail_template_timesheet_reminder_manager` are referring to different
models. The first one refers to `hr.employee`, while the latter refers a
`res.users`.

This causes the issue: the ID is used on a `res.users` while it should
be used on a `hr.employee`.

Closes #3314
opw-1894540

closes odoo/enterprise#3345