[FIX] account_accountant: fix uninstall issue
Commit b1d2aff10a66ea9a88e3af87f17d2260dd2597b4 undid the changes
made to account by the account_accountant module at uninstall,
in the safest possible way.

In this commit we are a bit more cavalier and remove all users from the
account_user group, so that they may not see that the module is the Odoo
reinvention of the Potemkin village.

Note that it also solves a bug resulting from the pair of commits
cfd4da43ee974a106e, 2955b8f1291ccd37.
Instead of renaming the menu, which breaks the translation, it moved the child
menuitems of account to the root menuitem of account_accountant.
Thus once the account_accountant module was installed, these submenus would be
orphan and appear on the main dashboard.
We give them back to their rightful (single) parent.

opw 1921183

closes odoo/enterprise#3352

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>