[FIX] web_enterprise: allow to check-in in mobile hr attendance
In mobile mode (mobile browser or app):
- Go to Attendances
- Click on the Check-in / Check-out button
Nothing happens.

The event `click .o_hr_attendance_sign_in_out_icon` is not triggered as
the click is captured by the `o_home_menu_background` which is added on
every `HrAttendance` template.

This was because the background overlay was added with an absolutely
positioned element on top of everything. This is not the desired
behavior as the overlay may interfere with static elements, and in this
case prevent to click on them. Using pointer-events: none would solve
the click problem but not the visual one, the overlay would still be
over the static elements.

opw-1916479
opw-1920269

closes odoo/enterprise#3369