[FIX] account_reports: extra options translatable
Commit e5cc535d4bd4f6d1ea7 improved the look of the search options, but
broke the translation of extra options. The only drawback of this
solution is the space between the option label and the comma.

Another solution would be to define these labels at the Python level and
re-use them in the template, but that's quite crappy.

Moreover, export missing terms of the module, especially following
commit https://github.com/odoo/odoo/commit/7f8631a913de7aebb9b95

opw-1908220

closes odoo/enterprise#3385