[FIX] hr_contract_salary: adapt tour to delay removal
This rev. is the counterpart of PR odoo/odoo#30106.

This rev. manually introduces a delay in the hr_contract_salary
tour where it is actually necessary, because we have to wait for
the jsSignature lib to render the signature (in a canvas), and the
lib does that in a setTimeout.

Before this rev., this tour passed by chance without selecting any
signature.

closes odoo/enterprise#3407