[FIX] l10n_be_reports: partner VAT listing: don't take discount lines for credit notes
Before that, making a discount by using a line with a negative quantity on the invoice resulted in the report considering this discount as a credit note, and thus always displaying the partner line. This was due to the fact credit notes are aggregated in the same subquery as the turnover, using the debit field of account move lines. To solve this, we need to split this subquery in two parts: one for the refund base, one for the turnover.

closes odoo/enterprise#3419