[FIX] web_grid: reload can set grid_anchor to undefined
In project forecast, accessing any field of the new line form of
the grid view crashes if the grid has been reloaded.

This is due to the reload function that assignes `undefined` to
'this.context.grid.anchor` if none of `old_context` or
`params.context` have a `grid_anchor` key set. This value is
ultimately used in `wrap_context` in `py_utils.js` where it crashes.

opw-1920489

closes odoo/enterprise#3461