[FIX] account_invoice_extract: add warning in log when OCR fails
When the OCR fails in Odoo there is only a status updated into the
field extract_state but there is nothing else shown. This means that
both the internal user and the partner have no idea what went wrong.
A minimal warning in the log would atleast improve the behaviour.

closes odoo/enterprise#3434