[FIX] web_studio: create related field to binaries (file/image)
Using studio, create a related field to a file or an image. The
creation raises a ValidationError due to missing char field.

The problem is due to the way files are created using studio, in
addition to create the binary field, it creates a char field for
the filename, this is the desired behavior for creating new normal
fields but not for a related fields.

closes odoo/enterprise#3435