[IMP] l10n_mx_edi: Make consistent the payment policy in the payment.
The attribute "MetodoPago" in the payment complement must be consistent
with the value in the invoice CFDI, not with the value in Odoo. For this
reason was improve that read the CFDI when verify if must be generated
the payment complement or not, because in some cases the CFDI could be
`PPD` and according to the method in the invoice must be `PUE`.

For this was added a new method that return the value in the attribute
`MetodoPago`

Also in the template is called the new method, to ensure that add the
value set in the CFDI from the invoice.

closes odoo/enterprise#3445