[IMP] l10n_mx_edi: Allow to force the REP generation
The Electronic Payment Receipt (REP) is required when the payment method
in the invoices related is `PPD`, but some customers require that in all
the payments generate the payment complement regardless the payment
method.

The label "Force payment complement" was added in the partner to
force the REF generation in all the payments that are made for it. And
if only is one exception, was added the button in the payment to force
the REF generation.

closes odoo/enterprise#3460